# Improve extension startup performance