--- a/package.json
+++ b/package.json
@@ -33,11 +33,14 @@
 	],
 	"main": "./out/extension.js",
 	"scripts": {
-		"vscode:prepublish": "npm run compile",
 		"compile": "tsc -p ./",
 		"lint": "tslint -p ./",
 		"watch": "tsc -watch -p ./",
-		"package": "vsce package"
+		"package": "vsce package",
+		"vscode:prepublish": "npm run esbuild-base -- --minify",
+		"esbuild-base": "esbuild ./src/extension.ts --bundle --outfile=out/extension.js --external:vscode --format=cjs --platform=node",
+		"esbuild": "npm run esbuild-base -- --sourcemap",
+		"esbuild-watch": "npm run esbuild-base -- --sourcemap --watch"
 	},
 	"contributes": {
 		"commands": [
@@ -392,9 +395,10 @@
 		"@types/prismjs": "^1.16.8",
 		"@types/vscode": "^1.68.0",
 		"@types/ws": "^8.2.2",
+		"esbuild": "^0.15.2",
 		"tslint": "^5.20.1",
 		"typescript": "^3.5.1",
-		"vsce": "^2.6.4"
+		"vsce": "^2.10.0"
 	},
 	"dependencies": {
 		"await-notify": "^1.0.1",