--- a/src/debugger/commands/command_parser.ts
+++ b/src/debugger/commands/command_parser.ts
@@ -134,9 +134,9 @@ export class CommandParser {
 						// we debugged that an exception occures during this.current_command.trigger(this.parameters)
 						// because we do not understand the root cause of the exception, we set the current command to undefined
 						// to avoid a infinite loop of parse_message(...)
-						this.current_command = undefined
+						this.current_command = undefined;
 						this.parameters = [];
-						console.log("Exception not catched. Reset current_command to avoid infinite loop.")
+						console.log("Exception not catched. Reset current_command to avoid infinite loop.");
 					}
 				} else {
 					this.current_command.param_count = this.parameters.shift();