Following the Microsoft documentation [here](https://code.visualstudio.com/api/working-with-extensions/bundling-extension#publishing), I rebuilt the extension's compilation pipeline to use a bundling tool called `esbuild`. This tool shoves all the javascript in the entire project into a single file, which sounds like a terrible idea until you look at this extension host profiler flame graph:

<details>
<summary>Details</summary>

![image](https://user-images.githubusercontent.com/18042232/184480722-273f879c-3cc9-4c97-a9cc-67df226c2c6c.png)

</details>

500+ms of almost nothing but module importing.

The highest startup time I saw tonight, on my machine was 650ms, and after bundling I'm averaging around 85ms. I have a user that's reported 1400ms startups, and the dev build I sent him cut that down to 630ms.

This cuts the compiled extension down from 796 files and 1.34MB, to 348 files and 295.19KB.

Additionally, this tool does some different linting, which revealed some warnings that I fixed. This extension now compiles with no warnings! (for now...)