@DaelonSuzuka 2022-08-13
I came back to add an edit saying that I didn't update the CI yet, but I decided to check it first:

![image](https://user-images.githubusercontent.com/18042232/184484160-f76dc76e-9b47-432c-bed0-7debb247931b.png)